refactor: moved rx-for into rxa folder
